refactor: extract SkillsSettings panel from SettingsDialog (#672)
Third (and largest) per-tab panel split. The skills tab (#629 / menu-config
#630) — catalog + enable/disable + reassign + reorder + import/remove/reload —
is self-contained on api.skills.* plus the pure menu-config helpers, so it
lifts whole into SkillsSettings.svelte. SettingsDialog 1912 → 1640 (and this is
the biggest single drop: −272). With it gone, the dialog also sheds the now-
unused menu-config / SKILL_MENUS / registerSkillInfos / skill-type imports and
the whole .skill-* CSS block.

App-wide sync preserved: every change still calls registerSkillInfos so the
command palette + slash commands track edits immediately (the panel did this
inline before; App.svelte already does the initial startup sync, so loading the
catalog lazily on tab-open rather than on dialog-open is equivalent).

Tests (8): components/SkillsSettings.test.ts mocks api.skills + the registry
sync and lets the real menu-config logic run — pins the on-mount catalog render
+ registry sync, the enable toggle + menu reassignment → setMenuConfig, Import
→ import+reload, Remove → remove(id), Reload, the import-error banner, and the
failed-to-load list.

Verified: pnpm lint (0 errors), full suite (2998 passed, +8), electron-forge
build + boot smoke.
